Passage float => double Optimisation de l'affichage des graphs utilisant des vecteurs de données GraphView class : ré-écriture / nettoyage + ajout de HD/SD rendering + affichage des axes avec uniquement les valeurs min/max possible Sauvegarde des résultats dans un fichier de données compressé, rechargé à l'ouverture du projet C14 : mémorisation de la dernière courbe choisie Vidage des corbeilles possible On quitte la fenêtre de calibration si une modification est enregistrée Classement des phases selon leur position Y dans la scène Correction des getThetaMin et Max : on utilise uniquement les events déjà initialisés (ce n'était pas le cas pour les phases...) Log HTML du modèle (à continuer...) Raw data affichées au-dessus de la densité "Help" ajouté pour les params MCMC et le Wiggle (dans les dialogues)

+++ src/mcmc/Generator.h
@@ -13,16 +13,16 @@ public:
     static int createSeed();
     static void initGenerator(const int seed);
     
-    static float randomUniform(float min = 0., float max = 1.);
-    static float gaussByDoubleExp(const float mean, const float sigma, const float min, const float max);
-    static float gaussByBoxMuller(const float mean, const float sigma);
+    static double randomUniform(double min = 0., double max = 1.);
+    static double gaussByDoubleExp(const double mean, const double sigma, const double min, const double max);
+    static double gaussByBoxMuller(const double mean, const double sigma);
     
 private:
     Generator(){}
-    static float boxMuller();
+    static double boxMuller();
     
     static std::mt19937 sGenerator;
-    static std::uniform_real_distribution<float> sDistribution;
+    static std::uniform_real_distribution<double> sDistribution;
 };
 
 #endif
