+++ src/plugins/plugin_gauss/PluginGauss.h
@@ -20,7 +20,7 @@ class DATATION_SHARED_EXPORT PluginGauss : public PluginAbstract
 public:
     PluginGauss();
     
-    float getLikelyhood(const float& t, const QJsonObject& data);
+    double getLikelyhood(const double& t, const QJsonObject& data);
     
     QString getName() const;
     QIcon getIcon() const;
